webui/requirements.txt
flask==3.1.*
psycopg2-binary==2.9.*
gunicorn==23.*
pyopenssl==24.*
